+++ atlasdb-tests-shared/src/main/java/com/palantir/atlasdb/sweep/AbstractSweepTaskRunnerTest.java
@@ -394,7 +394,7 @@ public abstract class AbstractSweepTaskRunnerTest {
                 wrappedCell("row", "c2"),
                 wrappedCell("row", "c3"),
                 wrappedCell("row", "c4"));
-        assertEquals(actualCells, expectedCells);
+        assertEquals(expectedCells, actualCells);
     }
 
     @Test